Moved FilteredList to address/util/collections.
