+++ src/main/java/address/util/collections/FilteredList.java
@@ -1,4 +1,4 @@
-package address.util;
+package address.util.collections;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ListChangeListener;
